Static casts

+++ aruw-mcb-project/src/aruwsrc/algorithms/plate_hit_tracker.cpp
@@ -45,14 +45,14 @@ void PlateHitTracker::update()
     {
         lastHitData.timestamp = newHitData.robotDataReceivedTimestamp;
 
-        lastHitData.hitAngle_chassisRelative_radians = Angle(lastHitData.plateID * M_PI / 2);
+        lastHitData.hitAngle_chassisRelative_radians = Angle(lastHitData.plateID * static_cast<float>(M_PI) / 2);
 
         lastHitData.hitAngle_worldRelative_radians = lastHitData.hitAngle_chassisRelative_radians +
                                                      transformer->getWorldToChassis().getYaw();
 
         // Update bins
         int binIndex = static_cast<int>(
-            lastHitData.hitAngle_worldRelative_radians.getWrappedValue() / (2 * M_PI / BIN_NUMBER));
+            lastHitData.hitAngle_worldRelative_radians.getWrappedValue() / (2 * static_cast<float>(M_PI) / BIN_NUMBER));
         // Add the hit to the bin
         // Magnitude is based on damage
         float damage = newHitData.receivedDps - lastHitData.lastDps;
