+++ aruw-mcb-project/src/aruwsrc/control/chassis/chassis_diagonal_drive_command.cpp
@@ -50,7 +50,7 @@ float ChassisDiagonalDriveCommand::computeAngleFromCenterForAutorotation(
         hypot(chassisVelocity[0][0], chassisVelocity[1][0]) > AUTOROTATION_DIAGONAL_SPEED)
     {
         angleFromCenterForChassisAutorotate =
-            WrappedFloat(turretAngleFromCenter, -M_PI_2, M_PI_2).getWrappedValue() + M_PI_4;
+            WrappedFloat(turretAngleFromCenter, -static_cast<float>(M_PI_2), static_cast<float>(M_PI_2)).getWrappedValue() + static_cast<float>(M_PI_4);
     }
     else
     {
