+++ aruw-mcb-project/src/aruwsrc/robot/sentry/sentry_agitator_constants.hpp
@@ -63,10 +63,10 @@ static constexpr aruwsrc::agitator::VelocityAgitatorSubsystemConfig AGITATOR_CON
      *
      * This should be positive or else weird behavior can occur
      */
-    .jammingVelocityDifference = M_TWOPI,
+    .jammingVelocityDifference = static_cast<float>(M_TWOPI),
     .jammingTime = 300,
     .jamLogicEnabled = true,
-    .velocityPIDFeedForwardGain = 500.0f / M_TWOPI,
+    .velocityPIDFeedForwardGain = 500.0f / static_cast<float>(M_TWOPI),
 };
 }  // namespace turretLeft
 
@@ -81,22 +81,22 @@ static constexpr aruwsrc::agitator::VelocityAgitatorSubsystemConfig AGITATOR_CON
      * The jamming constants. Agitator is considered jammed if difference between the velocity
      * setpoint and actual velocity is > jammingVelocityDifference for > jammingTime.
      */
-    .jammingVelocityDifference = 2.0f * M_TWOPI,
+    .jammingVelocityDifference = 2.0f * static_cast<float>(M_TWOPI),
     .jammingTime = 200,
     .jamLogicEnabled = true,
-    .velocityPIDFeedForwardGain = 500.0f / M_TWOPI,
+    .velocityPIDFeedForwardGain = 500.0f / static_cast<float>(M_TWOPI),
 };
 }  // namespace turretRight
 
 static constexpr tap::control::setpoint::MoveIntegralCommand::Config AGITATOR_ROTATE_CONFIG = {
     // magic numbers are fudge factors
-    .targetIntegralChange = M_TWOPI / AGITATOR_NUM_POCKETS - OVERSHOOT_FUDGE_FACTOR,
-    .desiredSetpoint = AGITATOR_MAX_ROF * (M_TWOPI / AGITATOR_NUM_POCKETS),
-    .integralSetpointTolerance = (M_TWOPI / AGITATOR_NUM_POCKETS) * 0.1f,
+    .targetIntegralChange = static_cast<float>(M_TWOPI) / AGITATOR_NUM_POCKETS - OVERSHOOT_FUDGE_FACTOR,
+    .desiredSetpoint = AGITATOR_MAX_ROF * (static_cast<float>(M_TWOPI) / AGITATOR_NUM_POCKETS),
+    .integralSetpointTolerance = (static_cast<float>(M_TWOPI) / AGITATOR_NUM_POCKETS) * 0.1f,
 };
 
-constexpr float UNJAM_VELOCITY = 0.35 * AGITATOR_MAX_ROF * (M_TWOPI / AGITATOR_NUM_POCKETS);
-constexpr float UNJAM_DISTANCE = 0.6f * (M_TWOPI / AGITATOR_NUM_POCKETS);
+constexpr float UNJAM_VELOCITY = 0.35f * AGITATOR_MAX_ROF * (static_cast<float>(M_TWOPI) / AGITATOR_NUM_POCKETS);
+constexpr float UNJAM_DISTANCE = 0.6f * (static_cast<float>(M_TWOPI) / AGITATOR_NUM_POCKETS);
 static constexpr aruwsrc::control::agitator::UnjamSpokeAgitatorCommand::Config
     AGITATOR_UNJAM_CONFIG = {
         .targetUnjamIntegralChange = UNJAM_DISTANCE,
